+++ pkg/analysis_server/lib/src/services/correction/dart/update_sdk_constraints.dart
@@ -32,7 +32,7 @@ class UpdateSdkConstraints extends CorrectionProducer {
         break;
       }
       pubspecFile = null;
-      folder = folder.parent;
+      folder = folder.parent2;
     }
     if (pubspecFile == null) {
       return;