+++ pkg/analyzer/test/file_system/overlay_file_system_test.dart
@@ -36,7 +36,7 @@ class FileTest extends OverlayTestSupport {
     File targetFile =
         provider.getFile(baseProvider.convertPath('/foo/test.dart'));
     expect(targetFile.exists, isFalse);
-    file.copyTo(file.parent!.parent!);
+    file.copyTo(file.parent2.parent2);
     expect(targetFile.exists, isTrue);
   }
 
@@ -46,7 +46,7 @@ class FileTest extends OverlayTestSupport {
     File targetFile =
         provider.getFile(baseProvider.convertPath('/foo/test.dart'));
     expect(targetFile.exists, isFalse);
-    file.copyTo(file.parent!.parent!);
+    file.copyTo(file.parent2.parent2);
     expect(targetFile.exists, isTrue);
     expect(targetFile.readAsStringSync(), 'overlay');
     provider.removeOverlay(targetFile.path);
@@ -59,7 +59,7 @@ class FileTest extends OverlayTestSupport {
     File targetFile =
         provider.getFile(baseProvider.convertPath('/foo/test.dart'));
     expect(targetFile.exists, isFalse);
-    file.copyTo(file.parent!.parent!);
+    file.copyTo(file.parent2.parent2);
     expect(targetFile.exists, isTrue);
     expect(targetFile.readAsStringSync(), 'overlay');
     provider.removeOverlay(targetFile.path);
@@ -164,7 +164,7 @@ class FileTest extends OverlayTestSupport {
   }
 
   test_parent() {
-    var parent = _file(exists: true).parent!;
+    var parent = _file(exists: true).parent2;
     expect(parent.exists, isTrue);
     expect(parent.path, defaultFolderPath);
   }
@@ -461,7 +461,7 @@ class FolderTest extends OverlayTestSupport {
         provider.getFolder(baseProvider.convertPath('/destination'));
 
     Folder copy = source.copyTo(destination);
-    expect(copy.parent, destination);
+    expect(copy.parent2, destination);
     _verifyStructure(copy, source);
   }
 
@@ -577,7 +577,7 @@ class FolderTest extends OverlayTestSupport {
 
   test_getChildren_nonExisting_withOverlay() {
     File file = _file(exists: false, withOverlay: true);
-    List<Resource> children = file.parent!.parent!.getChildren();
+    List<Resource> children = file.parent2.parent2.getChildren();
     expect(children, hasLength(1));
     expect(children[0], _isFolder);
   }
@@ -598,14 +598,15 @@ class FolderTest extends OverlayTestSupport {
   }
 
   test_parent_ofNonRoot() {
-    Folder parent = _folder(exists: true).parent!;
+    Folder parent = _folder(exists: true).parent2;
     expect(parent.exists, isTrue);
     expect(parent.path, baseProvider.convertPath('/foo'));
   }
 
   test_parent_ofRoot() {
-    var parent = _folder(exists: true, path: '/').parent;
-    expect(parent, isNull);
+    var parent = _folder(exists: true, path: '/').parent2;
+    expect(parent.exists, isTrue);
+    expect(parent.path, baseProvider.convertPath('/'));
   }
 
   @failingTest
@@ -734,7 +735,7 @@ class OverlayResourceProviderTest extends OverlayTestSupport {
 
   test_getFolder_notExisting_withOverlay() {
     File file = _file(exists: false, withOverlay: true);
-    Folder folder = file.parent!;
+    Folder folder = file.parent2;
     expect(folder, isNotNull);
     expect(folder.path, defaultFolderPath);
     expect(folder.exists, isTrue);