+++ pkg/analyzer/test/source/package_map_resolver_test.dart
@@ -44,7 +44,7 @@ class _PackageMapUriResolverTest {
     var b = provider.newFile(provider.convertPath('/bbb/b.dart'), '');
     expect(() {
       PackageMapUriResolver(provider, <String, List<Folder>>{
-        'pkg': <Folder>[a.parent!, b.parent!]
+        'pkg': <Folder>[a.parent2, b.parent2]
       });
     }, throwsArgumentError);
   }