+++ pkg/analyzer/CHANGELOG.md
@@ -2,6 +2,9 @@
 * Deprecated `TypeProvider.futureType2()`, `iterableType2()`, etc.
   Use corresponding `TypeProvider.futureType()`, `iterableType()`, etc.
 * Remove experimental markers from Null Safety APIs.
+* Added `Resource.parent2` and deprecated `Resource.parent`.
+* Added `Folder.isRoot`.
+* Added `FolderExtension` with `withAncestors`.
 
 ## 1.0.0
 * Stable null safety release.